fix(DragDrop): use preview when capturing drag & drop events
The reason is that the child component may consume the event (e.g.,
TextBox). In such case, our drag & drop will be cancelled when dragging
into the subview. Use the preview version of the events to workaround
the issue, because the preview version will reverse the event handling
order (from bubbling to tunneling). We should further inspect if there
is any side effect.