fix(dragDrop): Only handle the dragOver if the target view is droppable
Since we use the tunneling event routing, the parent view should not
handle the event if it cannot recognize. There is a known issue that if
the parent and the child are listening to a same event, only the parent
view can receive the dragOver. For the simplicity, I don't plan to add
more code logic for this case now. We could fix this issue only when we
do have this demand.